+++ packages/app-shell/src/PaymentMethod/components/StripeProvider.jsx
@@ -1,0 +1,35 @@
+import React, { useEffect, useState } from 'react';
+import {loadStripe} from '@stripe/stripe-js';
+import {
+  Elements,
+} from "@stripe/react-stripe-js";
+
+const StripeProvider = ({ children }) => {
+  const isLocal = !!window.location.host.match(/local/);
+  const STRIPE_PUBLISHABLE_KEY = isLocal ?
+    // test Stripe Publishable Key
+    'pk_test_CvOaedJTBPQLmI0YSnQsitzN' :
+    // production Stripe Publishable Key
+    'pk_dGKqAIFsUQonSYGPBM9Rek71IHOcL'
+    ;
+
+  const [stripe, setStripe] = useState(null)
+  useEffect(() => {
+    if (!stripe) {
+      setStripe(loadStripe(STRIPE_PUBLISHABLE_KEY))
+    }
+  }, [])
+
+  return (<Elements
+    stripe={stripe}
+    options={{
+      fonts: [{
+        cssSrc: 'https://fonts.googleapis.com/css?family=Roboto',
+      }],
+    }}
+  >
+    {children}
+  </Elements>)
+}
+
+export default StripeProvider
